requirements.txt
django
django-rest-framework
django-cors-headers
djangorestframework-simplejwt